Actualización: Definiciones de latas en archivo propio
