+++ sim/src/scenes.rs
@@ -1,117 +1,5 @@
+use crate::cans::*;
 use bevy::prelude::*;
-
-#[derive(Component)]
-pub struct CanData {
-    pos: Vec3,
-    rot: Quat,
-    size: CanSize,
-    texture: CanTexture,
-}
-
-pub enum CanSize {
-    Big,
-    Small,
-}
-
-pub enum CanTexture {
-    Shiny,
-    Dirty,
-}
-
-#[derive(Resource)]
-pub struct CanManager {
-    dirty_material_handle: Option<Handle<StandardMaterial>>,
-    shiny_material_handle: Option<Handle<StandardMaterial>>,
-    big_mesh_handle: Option<Handle<Mesh>>,
-    small_mesh_handle: Option<Handle<Mesh>>,
-}
-
-impl CanManager {
-    pub fn new() -> Self {
-        CanManager {
-            dirty_material_handle: None,
-            shiny_material_handle: None,
-            big_mesh_handle: None,
-            small_mesh_handle: None,
-        }
-    }
-
-    fn load_textures(
-        &mut self,
-        mut materials: ResMut<Assets<StandardMaterial>>,
-        asset_server: Res<AssetServer>,
-    ) {
-        self.dirty_material_handle = Some(materials.add(StandardMaterial {
-            base_color_texture: Some(asset_server.load("textures/black_01.png")),
-            reflectance: 0.3,
-            ..default()
-        }));
-
-        self.shiny_material_handle = Some(materials.add(StandardMaterial {
-            base_color_texture: Some(asset_server.load("textures/black_02.png")),
-            reflectance: 0.34,
-            ..default()
-        }));
-    }
-
-    fn load_meshes(&mut self, mut meshes: ResMut<Assets<Mesh>>) {
-        self.big_mesh_handle = Some(
-            meshes.add(
-                shape::Cylinder {
-                    radius: 0.15,
-                    height: 0.47,
-                    resolution: 15,
-                    segments: 1,
-                }
-                .into(),
-            ),
-        );
-        self.small_mesh_handle = Some(
-            meshes.add(
-                shape::Cylinder {
-                    radius: 0.15,
-                    height: 0.37,
-                    resolution: 15,
-                    segments: 1,
-                }
-                .into(),
-            ),
-        );
-    }
-
-    pub fn spawn(&mut self, commands: &mut Commands, can_data: CanData) {
-        let mesh = match can_data.size {
-            CanSize::Big => self.big_mesh_handle.clone().unwrap(),
-            CanSize::Small => self.small_mesh_handle.clone().unwrap(),
-        };
-
-        let material = match can_data.texture {
-            CanTexture::Shiny => self.shiny_material_handle.clone().unwrap(),
-            CanTexture::Dirty => self.dirty_material_handle.clone().unwrap(),
-        };
-
-        commands.spawn((
-            PbrBundle {
-                mesh,
-                material,
-                transform: Transform::from_xyz(can_data.pos.x, can_data.pos.y, can_data.pos.z)
-                    .with_rotation(can_data.rot),
-                ..default()
-            },
-            can_data,
-        ));
-    }
-}
-
-fn init_can_manager(
-    mut can_manager: ResMut<CanManager>,
-    meshes: ResMut<Assets<Mesh>>,
-    materials: ResMut<Assets<StandardMaterial>>,
-    asset_server: Res<AssetServer>,
-) {
-    can_manager.load_meshes(meshes);
-    can_manager.load_textures(materials, asset_server);
-}
 
 pub enum SceneName {
     Test,
@@ -292,7 +180,7 @@ fn spawn_basic_scene_with_cans(
     });
 
     commands.spawn((Camera3dBundle {
-        transform: Transform::from_xyz(0.01, 20.0, 0.0)
+        transform: Transform::from_xyz(4.01, 2.0, 10.0)
             .looking_at(Vec3::new(0.0, 0.0, 0.0), Vec3::Y),
         ..default()
     },));
